bhutan_document_processor/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.11.0
openai==2.1.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
